requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# Redis and Caching
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7
bcrypt==4.1.2

# API Documentation
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1

# Monitoring & Logging
prometheus-client==0.19.0
django-prometheus==2.3.1
structlog==23.2.0

# File Handling
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# Email
django-anymail==10.2
sendgrid==6.10.0

# Search
django-haystack==3.2.1
elasticsearch==8.11.0

# Task Queue
celery[redis]==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WebSocket
channels==4.0.0
channels-redis==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development
django-debug-toolbar==4.2.0
django-silk==5.0.4
ipython==8.17.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.1.0
